Move away the remaining model-related code to the model_manager
